refactor Sidebar into Presentational component
